feat: add book_editions table and split works from editions
